worker/requirements.txt
seestarpy
requests
python-dotenv
Pillow
